ci: add PR-time workflow (cargo check + npm build + clippy) (#55)
* ci: add PR-time workflow (cargo check + npm build + clippy)

Adds a PR-time CI workflow that catches compile/build errors before merge.

The previous setup had no CI on PRs — only `.github/workflows/release.yml`
ran on tag push (`v*`). The v2.6.0 release required 3 tag pushes / workflow
runs to ship because Rust compile errors and Tauri 2.11 version mismatches
were merged without being caught. This wastes 10–15 min per release cycle.

This workflow runs 3 parallel jobs on every PR to main/master:

  - frontend:     npm ci + npm run build           (Svelte/TS errors)
  - rust:         cargo check --all-targets        (type/borrow errors)
  - rust-clippy:  cargo clippy -- -D warnings      (lints as errors)

Each job has a 10-minute timeout. Cargo is cached with Swatinem/rust-cache,
npm with actions/setup-node cache: 'npm'. The release workflow is unchanged
and still handles bundling — this PR is just fast type-check feedback.

Mirrors Node 20, dtolnay/rust-toolchain@stable, and CARGO_INCREMENTAL=0
from release.yml. The rust jobs also run `npm ci && npm run build` first
because tauri::generate_context!() embeds build/ via include_dir! at
compile time, even during cargo check.

* ci: install Tauri Linux system deps (gtk3, webkit2gtk, etc.)

The first run of the PR-time CI failed with:

  The system library 'glib-2.0' required by crate 'glib-sys' was not found.

Tauri's Linux build pulls in glib-sys, webkit2gtk-sys, and other FFI
crates via the tauri + wry dependency tree. The build scripts invoke
pkg-config to locate the system libraries and headers, so the runner
needs the GTK 3 + webkit2gtk 4.1 dev packages installed.

release.yml doesn't have this problem because it only builds on
macOS/Windows — there's no Linux job there. The PR-time workflow
runs on ubuntu-latest, so the deps have to be installed explicitly.

Added an 'Install Linux system dependencies' step to both the
'rust' (cargo check) and 'rust-clippy' jobs, after the rust toolchain
setup and before the cargo cache. Mirrors the standard Tauri
prerequisites list (https://tauri.app/start/prerequisites/#linux).

* fix(clippy): address 7 lints surfaced by PR-time clippy job

The PR-time 'cargo clippy -D warnings' job surfaced 7 pre-existing
lints in the codebase. None of these are bugs — they're all clippy
style nits that the project has never run with -D warnings, so
they've been accumulating. Each one is addressed with the minimum
disruption possible:

- clippy::new_without_default (lib.rs:48)
  Added 'impl Default for AppState' that delegates to Self::new() so
  the existing log line in new() still fires.

- clippy::derivable_impls (config.rs:168)
  The AppConfig::default() impl uses 'default_*' helper functions to
  seed LoggingConfig with non-Default values. Added an allow with a
  comment explaining why the manual impl is intentional.

- clippy::redundant_pattern_matching (keychain.rs:51)
  Replaced 'matches!(entry.get_password(), Ok(_))' with
  'entry.get_password().is_ok()'.

- clippy::too_many_arguments (polling.rs:82)
  process_track has 8 args (threshold 7). Added an allow with a
  comment explaining why threading the per-poll mutable state through
  a single helper is intentional; a follow-up refactor to a
  PollContext struct is a separate concern.

- clippy::unnecessary_cast (polling.rs:243)
  Removed redundant 'as u64' cast on expiry_buffer_seconds (already
  u64).

- clippy::let_unit_value (menu.rs:126)
  The 'let _ = app_handle.exit(0);' idiom is intentional — it
  suppresses the unused_must_use lint on a cross-platform call that
  returns Result on Unix/macOS and () on Windows. Wrapped the
  expression in a block with '#[allow(clippy::let_unit_value)]'.

- clippy::needless_borrow (commands.rs:679)
  Removed the redundant '&' before 'state.inner()' in
  'Arc::clone(&state.inner())'.

No behaviour change. Just lint hygiene so PR-time CI stays green.

* ci: harden workflow per CodeRabbit review

CodeRabbit flagged two security/hardening concerns on the workflow:

1. persist-credentials on checkout
   actions/checkout persists the GITHUB_TOKEN in the runner's local
   git config by default. That token can be exfiltrated by later
   steps (or by an untrusted post-action) if the workflow ever runs
   against untrusted input. Setting persist-credentials: false on
   each of the 3 checkout steps prevents the token from being
   written to .git/config unless an explicit step later opts in.

2. SHA-pin all 'uses:' references
   Tag-pinned actions like actions/checkout@v4 are mutable — if a
   tag is deleted/recreated or an action is compromised, the
   workflow could pull a different commit. Pinning to the immutable
   commit SHA makes the supply chain deterministic; the trailing
   version comment (e.g. '# v4.3.1') is a hint for humans and
   dependabot updates.

Pinned SHAs (latest as of 2026-06-09):
- actions/checkout        @ 34e114876b0b11c390a56381ad16ebd13914f8d5  # v4.3.1
- actions/setup-node      @ 49933ea5288caeca8642d1e84afbd3f7d6820020  # v4.4.0
- dtolnay/rust-toolchain  @ bd41891a8e7f4b8649f6d684415e1a6155fe4e22  # 1.83
- Swatinem/rust-cache     @ c19371144df3bb44fab255c43d04cbc2ab54d1c4  # v2.9.1

dtolnay/rust-toolchain switched from @stable to @1.83 — same Rust
stable that release.yml was using via @stable on the last release.
The project's Cargo.toml doesn't pin a specific MSRV, so pinning
the action to 1.83 matches what 'stable' would have resolved to
during the v2.6.0 release build. Bumping the SHA is the project's
explicit decision to upgrade the toolchain.

No behaviour change in the run; just removes two real attack
surfaces.

* ci: pin rust-toolchain to 1.96 (latest released stable)

The 1.83 pin in the previous commit was too old — the project's
dependency dlopen2 v0.8.2 requires 'edition2024', which was
stabilized in Rust 1.85.

The 1.100 pin in my first follow-up was also wrong — dtolnay/
rust-toolchain's branch list goes up to 1.100, but those are
forward-looking branches. rust-lang.org only has releases up to
1.96.0 on the CDN (verified: rust-1.97.0 through rust-1.100.0
all return 404 from static.rust-lang.org/dist/). The action
tried to install 1.100.0 and failed with a 404.

1.96.0 is what @stable would have resolved to during the v2.6.0
release build (which succeeded). Pinning to the 1.96 branch SHA
preserves the v2.6.0 build baseline without silent drift from
'whatever stable is today'.

Bumping the SHA is now the project's explicit decision to upgrade
Rust.

---------

+++ src-tauri/src/polling.rs
@@ -79,6 +79,13 @@ fn with_jitter(base_secs: u64) -> u64 {
     (base_secs as f64 + jitter).max(1.0) as u64
 }
 
+// process_track is intentionally a single helper that owns the per-poll
+// mutable state in one place. Threading the same fields through every
+// call site would either duplicate them across the polling loop or
+// require a new "PollContext" struct that mirrors the existing parameter
+// list. The signature has grown because the polling loop has grown;
+// refactoring is a separate concern from PR-time CI hygiene.
+#[allow(clippy::too_many_arguments)]
 fn process_track(
     app: &AppHandle,
     state: &Arc<AppState>,
@@ -240,7 +247,7 @@ fn process_track(
             let remaining_ms = track.duration_ms.saturating_sub(corrected_progress_ms);
             let buffer_ms = config
                 .as_ref()
-                .map(|c| c.polling.expiry_buffer_seconds as u64)
+                .map(|c| c.polling.expiry_buffer_seconds)
                 .unwrap_or(10)
                 * 1000;
             let expiry =
